move rawmetadata to new file
