+++ warehouse/query-core/src/test/java/datawave/query/GroupedNormalFlattenQueryTest.java
@@ -6,7 +6,6 @@ import datawave.query.exceptions.FullTableScansDisallowedException;
 import datawave.query.testframework.AbstractFields;
 import datawave.query.testframework.AbstractFunctionalQuery;
 import datawave.query.testframework.AccumuloSetupHelper;
-import datawave.query.testframework.BaseRawData;
 import datawave.query.testframework.DataTypeHadoopConfig;
 import datawave.query.testframework.FieldConfig;
 import datawave.query.testframework.FileLoaderFactory;
@@ -14,6 +13,7 @@ import datawave.query.testframework.FlattenData;
 import datawave.query.testframework.FlattenDataType;
 import datawave.query.testframework.FlattenDataType.FlattenBaseFields;
 import datawave.query.testframework.RawDataManager;
+import datawave.query.testframework.RawMetaData;
 import org.apache.log4j.Logger;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -224,11 +224,11 @@ public class GroupedNormalFlattenQueryTest extends AbstractFunctionalQuery {
             headers = Stream.of(GroupedNormalField.values()).map(e -> e.name()).collect(Collectors.toList());
         }
         
-        static final Map<String,BaseRawData.RawMetaData> metadataMapping = new HashMap<>();
+        static final Map<String,RawMetaData> metadataMapping = new HashMap<>();
         
         static {
             for (GroupedNormalField field : GroupedNormalField.values()) {
-                BaseRawData.RawMetaData data = new BaseRawData.RawMetaData(field.name(), field.normalizer, false);
+                RawMetaData data = new RawMetaData(field.name(), field.normalizer, false);
                 metadataMapping.put(field.name().toLowerCase(), data);
             }
         }
